backend/requirements.txt
fastapi
uvicorn
google-generativeai
langchain
langchain-google-genai
sqlalchemy
psycopg2-binary
ipfshttpclient
web3
pydantic
python-dotenv